fix: Force interiorId refresh on json load
`interiorId` can change between sessions, especially on two separate clients. When we were loading from the json files, we previously assumed that the interiorId would be correct when present. now we ensure that it isn't present any time the file is loaded from json and regrab it for the current session.